backend/baserv/requirements.txt
asgiref==3.5.2
autopep8==1.7.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
distlib==0.3.6
Django==4.1.2
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
filelock==3.8.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
openapi-codec==1.3.2
packaging==21.3
pipenv==2022.10.25
platformdirs==2.5.2
pycodestyle==2.9.1
pycparser==2.21
PyJWT==2.6.0
PyMySQL==1.0.2
pyparsing==3.0.9
pytz==2022.5
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.17.6
sqlparse==0.4.3
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.16.6
virtualenv-clone==0.5.7
whitenoise==6.2.0
